% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/float_image.R
\name{float_image}
\alias{float_image}
\title{Produces HTML <div> Tag For Float Images}
\usage{
float_image(src = NULL, float = "right", width = "200")
}
\arguments{
\item{src}{source}

\item{float}{float direction}

\item{width}{width of image}
}
\description{
Produces HTML <div> Tag For Float Images.
}
\examples{
float_image("/images/googleformr_logo.png")
}

