% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gist.R
\name{gist}
\alias{gist}
\title{Create Script to Embed Gist}
\usage{
gist(url)
}
\arguments{
\item{url}{the gist url}
}
\description{
Create Script to Embed Gist.
}
\examples{
gist("https://gist.github.com/nisrulz/11c0d63428b108f10c83")
}

