% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_star.R
\name{github_star}
\alias{github_star}
\title{Interactively Create Github Buttons}
\usage{
github_star(repo, acct = "data-steve", size = "mega", script = FALSE)
}
\arguments{
\item{repo}{github repo}

\item{acct}{github acct}

\item{size}{button size}

\item{script}{include github-button.js or not}
}
\description{
Interactively Create Github Buttons.
}
\examples{
\dontrun{
github_star("blogtools", "data-steve")
}
}

