% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links_liked.R
\name{links_liked}
\alias{links_liked}
\title{Create HTML for Links I Liked Posts}
\usage{
links_liked(text, url, shrtn = FALSE, hashtags = NULL,
  acct = "data_steve", script = FALSE)
}
\arguments{
\item{text}{text to be shown over hyperlink}

\item{url}{url for link}

\item{shrtn}{TRUE/FALSE whether to shorten url}

\item{hashtags}{single string of hashtags quoted and comma-separated}

\item{acct}{twitter acct}

\item{script}{T/F for whether to include twitter script}
}
\description{
Create HTML for Links I Liked Posts.
}
\examples{
links_liked(text = "Using d3 to 'discover' koala's in nature"
            , url = "http://www.koalastothemax.com/"
            , shrtn=TRUE
            , hashtags="d3js,koalafield,WWF")
}

