% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_blank.R
\name{target_blank}
\alias{target_blank}
\title{Generate Target Blank <a>}
\usage{
target_blank(text, href)
}
\arguments{
\item{text}{text}

\item{href}{link}
}
\description{
Generate Target Blank <a>.
}
\examples{
target_blank("Google Search", "https://www.google.com/")
}

