% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{tweet}
\alias{tweet}
\title{Interactively Create Button to Embed Tweets}
\usage{
tweet(acct, hashtags, size = "large")
}
\arguments{
\item{acct}{twitter acct}

\item{hashtags}{tags to include in tweet}

\item{size}{size of button}
}
\description{
Interactively Create Button to Embed Tweets.
}
\examples{
\dontrun{
tweet("data_steve", c("R", "Data Science"))
}
}

