% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_isgd.R
\name{url_isgd}
\alias{url_isgd}
\title{Shortens or Expands URLs}
\usage{
url_isgd(url, shrtn = TRUE)
}
\arguments{
\item{url}{url to be shortened or expanded}

\item{shrtn}{T/F whether to shorten or expand}
}
\description{
Shortens or Expands URLs. It uses the is.gd API : https://is.gd/.
Once URL has been shortened initially, same shortened url is returned
}
\examples{
url_isgd("http://andrewgelman.com/2016/03/22/text-sentiment-visualizer/")
}

