% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{tweet}
\alias{tweet}
\title{Interactively Create Button to Embed Tweets}
\usage{
tweet(hashtags, acct = "data_steve", size = "large")
}
\arguments{
\item{hashtags}{tags to include in tweet}

\item{acct}{twitter acct}

\item{size}{size of button}
}
\description{
Interactively Create Button to Embed Tweets.
}
\examples{
\dontrun{
tweet("rstats, DataScience")
}
}

